Move RtcRtpEncodingParameters into separate module
